pyproject.toml
[tool.poetry]
name = "core"
version = "2.2.42"
description = ""
authors = ["Indoc Research"]

[tool.poetry.dependencies]
python = ">=3.7,<3.11"
fastapi = "0.88.0"
fastapi-utils = "0.2.1"
aniso8601 = "8.0.0"
attrs = "19.3.0"
certifi = "2020.4.5.1"
chardet = "3.0.4"
itsdangerous = "1.1.0"
Jinja2 = "2.11.3"
jsonschema = "3.2.0"
MarkupSafe = "1.1.1"
PyJWT = "2.6.0"
python-json-logger = "2.0.2"
pytz = "2020.1"
requests = "2.26.0"
toml = "0.10.1"
Werkzeug = "1.0.1"
wrapt = "1.12.1"
pydantic = "1.10.5"
asyncpg = "0.27.0"
opentelemetry-exporter-otlp = "1.6.2"
opentelemetry-instrumentation = "^0.30b1"
opentelemetry-instrumentation-fastapi = "^0.30b1"
opentelemetry-instrumentation-httpx = "^0.30b1"
uvicorn = "0.17.6"
httpx = "0.23.0"
pilot-platform-common = "0.3.0"
redis = "4.5.5"
pact-python = "^1.6.0"
protobuf = "3.20.3"

[tool.poetry.dev-dependencies]
pytest = "7.1.2"
pytest-cov = "3.0.0"
pytest-mock = "3.7.0"
pytest-httpx = "0.21.0"
pytest-random-order = "1.1.1"
requests-mock = "1.9.3"
testcontainers = "3.6.1"
trio = "0.21.0"
async-asgi-testclient = "^1.4.9"
pytest-asyncio = "^0.17.2"
Faker = "16.8.1"

[tool.pytest.ini_options]
addopts = "-v --random-order"
testpaths = "tests"
asyncio_mode = "auto"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
